refactor(network-interception): rename package from network-policy (1/8)
Use @packages/network-interception from the start of the stacked refactor
and add README documenting the ports/adapters architecture boundary.
