+++ .circleci/scripts/generate-pipeline-parameters.sh
@@ -158,7 +158,7 @@ while IFS= read -r file; do
       driver_tests=true
       system_tests=true
       ;;
-    packages/network-policy/*)
+    packages/network-interception/*)
       server_tests=true
       driver_tests=true
       system_tests=true
