+++ knip.json
@@ -414,7 +414,7 @@
         "lib/**/*.{ts,tsx,js,jsx}"
       ]
     },
-    "packages/network-policy": {
+    "packages/network-interception": {
       "entry": [
         "lib/index.ts"
       ],
@@ -799,17 +799,17 @@
       "exports",
       "types"
     ],
-    "packages/network-policy/lib/types/external-types.ts": [
+    "packages/network-interception/lib/types/external-types.ts": [
       "exports",
       "types"
     ],
-    "packages/network-policy/lib/ports/driven-ports.ts": [
+    "packages/network-interception/lib/ports/driven-ports.ts": [
       "types"
     ],
-    "packages/network-policy/lib/ports/driving-ports.ts": [
+    "packages/network-interception/lib/ports/driving-ports.ts": [
       "types"
     ],
-    "packages/network-policy/lib/exchange/network-exchange.ts": [
+    "packages/network-interception/lib/exchange/network-exchange.ts": [
       "types"
     ],
     "npm/react/cypress/component/advanced/lazy-loaded-suspense/Dog.tsx": [
