+++ packages/network-interception/.eslintignore
