+++ packages/network-interception/lib/types/external-types.ts
