+++ packages/network-interception/lib/types/internal-types.ts
