+++ packages/network-interception/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@packages/network-policy",
+  "name": "@packages/network-interception",
   "version": "0.0.0-development",
   "private": true,
   "main": "./lib/index",
