+++ packages/network-interception/tsconfig.json
